+++ tests/StaticTests.php
@@ -1,5 +1,8 @@
 <?php namespace Cviebrock\EloquentTaggable\Test;
 
+use Cviebrock\EloquentTaggable\Services\TagService;
+
+
 /**
  * Class StaticTests
  */
@@ -24,7 +27,7 @@ class StaticTests extends TestCase
     /**
      * @inheritdoc
      */
-    public function setUp()
+    public function setUp(): void
     {
         parent::setUp();
 
@@ -50,7 +53,7 @@ class StaticTests extends TestCase
     /**
      * Test finding all the tags for a model, in a list.
      */
-    public function testAllTagsList()
+    public function testAllTagsList(): void
     {
         $tags = TestModel::allTagsList();
 
@@ -60,7 +63,7 @@ class StaticTests extends TestCase
     /**
      * Test renaming the tags for a model.
      */
-    public function testRenameTags()
+    public function testRenameTags(): void
     {
         TestModel::renameTag('Apple', 'Apricot');
 
@@ -76,7 +79,7 @@ class StaticTests extends TestCase
     /**
      * Test getting the popular tags for a model.
      */
-    public function testPopularTags()
+    public function testPopularTags(): void
     {
         $tags = TestModel::popularTags();
         $expected = [
@@ -91,7 +94,7 @@ class StaticTests extends TestCase
     /**
      * Test getting the popular tags for a model, normalized.
      */
-    public function testPopularTagsNormalized()
+    public function testPopularTagsNormalized(): void
     {
         $tags = TestModel::popularTagsNormalized();
         $expected = [
@@ -106,7 +109,7 @@ class StaticTests extends TestCase
     /**
      * Test getting the popular tags for a model, with a limit.
      */
-    public function testPopularTagsLimited()
+    public function testPopularTagsLimited(): void
     {
         $tags = TestModel::popularTags(2);
 
@@ -121,13 +124,30 @@ class StaticTests extends TestCase
     /**
      * Test getting the popular tags for a model, with a limit.
      */
-    public function testPopularTagsLimitedNormalized()
+    public function testPopularTagsLimitedNormalized(): void
     {
         $tags = TestModel::popularTagsNormalized(2);
 
         $expected = [
             'apple'  => 3,
             'cherry' => 2,
+        ];
+
+        $this->assertArrayValuesAreEqual($expected, $tags);
+    }
+
+    public function testAllTagsWithCount()
+    {
+        $tags = app(TagService::class)
+            ->getPopularTags(null, null, 0)
+            ->pluck('taggable_count', 'normalized')
+            ->toArray();
+
+        $expected = [
+            'apple'  => 4,
+            'cherry' => 2,
+            'banana' => 1,
+            'durian' => 1,
         ];
 
         $this->assertArrayValuesAreEqual($expected, $tags);